Import "otis", an HTTP request rewriting engine
Otis is an HTTP request rewriting engine in Erlang.

The configuration file is written in YAML and uses variables to read or
change each part of the request. It also uses keywords to describe the
conditions and the actions. Beside strings, the format understands
various types such as integer or IP address.

An Erlang module is then generated based on this configuration file.
Therefore the final engine is specifically optimized for the usage.